+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 Change Log
 ==========
 
+Version 1.2.1
+-------------
+
+  * Added `mutable` attribute to `@Decor` to allow for mutable decorated objeect.
+
 Version 1.2.0
 -------------
 
-  * Added '@Factory' annotation to generate an interface that create instance of the marked interface.
+  * Added `@Factory` annotation to generate an interface that create instance of the marked interface.
 
 Version 1.1.1
 -------------
@@ -14,12 +19,12 @@ Version 1.1.1
 Version 1.1.0
 -------------
 
-  * Both '@Noop' and '@Decor' support specifying the name of the generated class as the optional value.
+  * Both `@Noop` and `@Decor` support specifying the name of the generated class as the optional value.
 
 Version 1.0.0
 -------------
 
-  * Added '@Decor' annotation to generate a decorator implementation of the marked interface.
+  * Added `@Decor` annotation to generate a decorator implementation of the marked interface.
 
 
 Version 0.1.0 *(unpublished)*